refactor: Update shader function to use uint for instanceId
- Changed the type of `instanceId` in the shader function from `float` to `uint`, enhancing type consistency and safety across shader implementations.
- This update aligns with recent changes to other shader properties, improving overall code clarity and maintainability.

+++ src/foundation/definitions/ShaderSemantics.ts
@@ -657,7 +657,7 @@ function getShaderPropertyOfGlobalDataRepository(_engine: Engine, info: ShaderSe
 
   if (!isTexture) {
     funcDef = `
-  ${returnType} get_${info.semantic}(float instanceId, int index) {
+  ${returnType} get_${info.semantic}(uint instanceId, uint index) {
     ${str}
   }
 `;
